refactor: 자동차 갯수를 확인하는 기능을 ValidateUtils가 아닌 Cars가 맡도록 변경
