+++ src/main/java/racingCar/utils/ValidateUtils.java
@@ -5,26 +5,18 @@ import racingCar.exception.InvalidNumOfMoveException;
 
 public class ValidateUtils {
 
-    public static final int MIN_NUM_OF_CARS = 2;
+
     public static final int MIN_NAME_LENGTH = 1;
     public static final int MAX_NAME_LENGTH = 5;
-    public static final String ERROR_MESSAGE_CAR_NAMES_COUNT = "자동차 이름은 1개 이상이어야 합니다.";
+
     public static final String ERROR_MESSAGE_CAR_NAME_NULL_OR_EMPTY = "자동차 이름은 공백이 안됩니다.";
     private static final String ERROR_MESSAGE_ROUND_COUNT_NUMBER = "시도 횟수는 숫자이어야 합니다.";
 
-
-    public static String[] validateNumOfCarNames(String[] carNames) {
-        if (carNames.length < MIN_NUM_OF_CARS) {
-            throw new InvalidCarNameException(ERROR_MESSAGE_CAR_NAMES_COUNT);
-        }
-        return carNames;
-    }
-
     public static boolean validateCarName(String carName) {
-        return validateCarNameLength(carName) && validateCarNameNotEmpty(carName);
+        return isProperLength(carName) && isNotBlank(carName);
     }
 
-    private static boolean validateCarNameLength(String carName) {
+    private static boolean isProperLength(String carName) {
         try {
             return carName.length() >= MIN_NAME_LENGTH && carName.length() <= MAX_NAME_LENGTH;
         } catch (NullPointerException exception) {
@@ -32,7 +24,7 @@ public class ValidateUtils {
         }
     }
 
-    private static boolean validateCarNameNotEmpty(String carName) {
+    private static boolean isNotBlank(String carName) {
         return !carName.isEmpty();
     }
